新增删除ftp图片功能

+++ src/main/java/com/caizhixiang/springboot/service/impl/ImageServiceImpl.java
@@ -1,15 +1,18 @@
 package com.caizhixiang.springboot.service.impl;
 
+import com.caizhixiang.springboot.ftp.FtpClient;
 import com.caizhixiang.springboot.mapper.ImageMapper;
 import com.caizhixiang.springboot.mapper.entity.Image;
 import com.caizhixiang.springboot.service.ImageService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tk.mybatis.mapper.entity.Example;
 
+import java.io.IOException;
 import java.util.Date;
 import java.util.List;
 
@@ -24,6 +27,8 @@ import java.util.List;
 public class ImageServiceImpl implements ImageService {
     @Autowired
     private ImageMapper mapper;
+    @Autowired
+    private FtpClient ftpClient;
 
     @Override
     public List<Image> findAll() {
@@ -53,7 +58,20 @@ public class ImageServiceImpl implements ImageService {
 
     @Override
     public void remove(Integer id) {
-        mapper.deleteByPrimaryKey(id);
+        Image image = mapper.selectByPrimaryKey(id);
+        if (image != null) {
+            mapper.deleteByPrimaryKey(id);
+            String url = image.getUrl();
+            if (StringUtils.isNotBlank(url)) {
+                url = url.substring(url.lastIndexOf("/") + 1);
+                try {
+                    ftpClient.deleteFile(url);
+                } catch (IOException e) {
+                    log.error("删除文件失败");
+                    e.printStackTrace();
+                }
+            }
+        }
     }
 
     @Override
